Avoid endless loop in uffs_ftruncate()
If a file spans more than one block and that block isn't found during
mount it would be possible for uffs_ftruncate() to end up in an endless
loop. Ensure that this doesn't happen by breaking the loop if
do_TruncateInternalWithBlockRecover() fails for any reason.

The error code is already set by do_TruncateInternalWithBlockRecover().